+++ lib/searchkick/railtie.rb
@@ -1,4 +1,4 @@
-module Searckick
+module Searchkick
   class Railtie < Rails::Railtie
     rake_tasks do
       load "tasks/searchkick.rake"