+++ test/boost_test.rb
@@ -66,7 +66,7 @@ class BoostTest < Minitest::Test
   def test_boost_by_missing
     store [
       {name: "Tomato A"},
-      {name: "Tomato B", orders_count: 10},
+      {name: "Tomato B", orders_count: 10}
     ]
 
     assert_order "tomato", ["Tomato A", "Tomato B"], boost_by: {orders_count: {missing: 100}}
@@ -166,7 +166,7 @@ class BoostTest < Minitest::Test
       {name: "San Marino", latitude: 43.9333, longitude: 12.4667, found_rate: 0.2}
     ]
 
-    assert_order "san", ["San Antonio","San Francisco", "San Marino"], boost_by: {found_rate: {factor: 100}}, boost_by_distance: {location: {origin: [37, -122], scale: "1000mi"}}
+    assert_order "san", ["San Antonio", "San Francisco", "San Marino"], boost_by: {found_rate: {factor: 100}}, boost_by_distance: {location: {origin: [37, -122], scale: "1000mi"}}
     assert_order "san", ["San Francisco", "San Antonio", "San Marino"], boost_by: {found_rate: {factor: 100}}, boost_by_distance: {location: {origin: [37, -122], scale: "1000mi", factor: 100}}
   end