+++ test/order_test.rb
@@ -27,7 +27,7 @@ class OrderTest < Minitest::Test
   end
 
   def test_unmapped_type
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
     assert_order "product", [], order: {not_mapped: {unmapped_type: "long"}}
     assert_order_relation [], Product.search("product").order(not_mapped: {unmapped_type: "long"})
   end
@@ -36,5 +36,12 @@ class OrderTest < Minitest::Test
     store [{name: "San Francisco", latitude: 37.7833, longitude: -122.4167}]
     assert_order "francisco", ["San Francisco"], order: [{_geo_distance: {location: "0,0"}}]
     assert_order_relation ["San Francisco"], Product.search("francisco").order([{_geo_distance: {location: "0,0"}}])
+  end
+
+  def test_script
+    store_names ["Red", "Green", "Blue"]
+    order = {_script: {type: "number", script: {source: "doc['name'].value.length() * -1"}}}
+    assert_order "*", ["Green", "Blue", "Red"], order: order
+    assert_order_relation ["Green", "Blue", "Red"], Product.search("*").order(order)
   end
 end