+++ test/reindex_test.rb
@@ -14,7 +14,7 @@ class ReindexTest < Minitest::Test
 
     product = Product.find_by!(name: "Product A")
     product.destroy
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
     assert_equal true, product.reindex
   end
 
@@ -25,25 +25,25 @@ class ReindexTest < Minitest::Test
     perform_enqueued_jobs do
       assert_equal true, product.reindex(mode: :async)
     end
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
     assert_search "product", ["Product A"]
   end
 
   def test_record_queue
-    reindex_queue = Product.search_index.reindex_queue
+    reindex_queue = Product.searchkick_index.reindex_queue
     reindex_queue.clear
 
     store_names ["Product A", "Product B"], reindex: false
 
     product = Product.find_by!(name: "Product A")
     assert_equal true, product.reindex(mode: :queue)
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
     assert_search "product", []
 
     perform_enqueued_jobs do
       Searchkick::ProcessQueueJob.perform_now(class_name: "Product")
     end
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
     assert_search "product", ["Product A"]
   end
 
@@ -51,7 +51,7 @@ class ReindexTest < Minitest::Test
     store_names ["Product A", "Product B"], reindex: false
 
     product = Product.find_by!(name: "Product A")
-    assert_equal true, Product.search_index.reindex([product], refresh: true)
+    assert_equal true, Product.searchkick_index.reindex([product], refresh: true)
     assert_search "product", ["Product A"]
   end
 
@@ -106,7 +106,7 @@ class ReindexTest < Minitest::Test
       Product.find_by(name: "Product B").update!(name: "DO NOT INDEX")
     end
     assert_equal true, Product.where(name: "DO NOT INDEX").reindex
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
     assert_search "product", ["Product A"]
   end
 
@@ -116,7 +116,7 @@ class ReindexTest < Minitest::Test
     perform_enqueued_jobs do
       Product.where(name: "Product B").reindex(mode: :async)
     end
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
     assert_search "product", ["Product A", "Product B"]
   end
 
@@ -128,32 +128,75 @@ class ReindexTest < Minitest::Test
     perform_enqueued_jobs do
       assert_equal true, Product.where(name: "DO NOT INDEX").reindex(mode: :async)
     end
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
     assert_search "product", ["Product A"]
   end
 
+  def test_relation_async_routing
+    store_names ["Store A"], Store, reindex: false
+    perform_enqueued_jobs do
+      Store.where(name: "Store A").reindex(mode: :async)
+    end
+    Store.searchkick_index.refresh
+    assert_search "*", ["Store A"], {routing: "Store A"}, Store
+  end
+
   def test_relation_queue
-    reindex_queue = Product.search_index.reindex_queue
+    reindex_queue = Product.searchkick_index.reindex_queue
     reindex_queue.clear
 
     store_names ["Product A"]
     store_names ["Product B", "Product C"], reindex: false
 
     Product.where(name: "Product B").reindex(mode: :queue)
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
     assert_search "product", ["Product A"]
 
     perform_enqueued_jobs do
       Searchkick::ProcessQueueJob.perform_now(class_name: "Product")
     end
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
     assert_search "product", ["Product A", "Product B"]
+  end
+
+  def test_relation_queue_all
+    reindex_queue = Product.searchkick_index.reindex_queue
+    reindex_queue.clear
+
+    store_names ["Product A"]
+    store_names ["Product B", "Product C"], reindex: false
+
+    Product.all.reindex(mode: :queue)
+    Product.searchkick_index.refresh
+    assert_search "product", ["Product A"]
+
+    perform_enqueued_jobs do
+      Searchkick::ProcessQueueJob.perform_now(class_name: "Product")
+    end
+    Product.searchkick_index.refresh
+    assert_search "product", ["Product A", "Product B", "Product C"]
+  end
+
+  def test_relation_queue_routing
+    reindex_queue = Store.searchkick_index.reindex_queue
+    reindex_queue.clear
+
+    store_names ["Store A"], Store, reindex: false
+    Store.where(name: "Store A").reindex(mode: :queue)
+    Store.searchkick_index.refresh
+    assert_search "*", [], {}, Store
+
+    perform_enqueued_jobs do
+      Searchkick::ProcessQueueJob.perform_now(class_name: "Store")
+    end
+    Store.searchkick_index.refresh
+    assert_search "*", ["Store A"], {routing: "Store A"}, Store
   end
 
   def test_relation_index
     store_names ["Product A"]
     store_names ["Product B", "Product C"], reindex: false
-    Product.search_index.reindex(Product.where(name: "Product B"), refresh: true)
+    Product.searchkick_index.reindex(Product.where(name: "Product B"), refresh: true)
     assert_search "product", ["Product A", "Product B"]
   end
 
@@ -169,9 +212,11 @@ class ReindexTest < Minitest::Test
     index.refresh
     assert_equal 1, index.total_docs
 
-    assert Searchkick.reindex_status(reindex[:name])
+    reindex_status = Searchkick.reindex_status(reindex[:name])
+    assert_equal true, reindex_status[:completed]
+    assert_equal 0, reindex_status[:batches_left]
 
-    Product.search_index.promote(reindex[:index_name])
+    Product.searchkick_index.promote(reindex[:index_name])
     assert_search "product", ["Product A"]
   end
 
@@ -189,12 +234,12 @@ class ReindexTest < Minitest::Test
   end
 
   def test_full_async_wait
-    skip "Need to fix for test adapter"
-
     store_names ["Product A"], reindex: false
 
-    capture_io do
-      Product.reindex(mode: :async, wait: true)
+    perform_enqueued_jobs do
+      capture_io do
+        Product.reindex(mode: :async, wait: true)
+      end
     end
 
     assert_search "product", ["Product A"]
@@ -216,15 +261,21 @@ class ReindexTest < Minitest::Test
     Sku.destroy_all
   end
 
+  def test_full_queue
+    assert_warns "Full reindex does not support :queue mode - use :async mode instead" do
+      Product.reindex(mode: :queue)
+    end
+  end
+
   def test_full_refresh_interval
     reindex = Product.reindex(refresh_interval: "30s", mode: :async, import: false)
     index = Searchkick::Index.new(reindex[:index_name])
-    assert_nil Product.search_index.refresh_interval
+    assert_nil Product.searchkick_index.refresh_interval
     assert_equal "30s", index.refresh_interval
 
-    Product.search_index.promote(index.name, update_refresh_interval: true)
+    Product.searchkick_index.promote(index.name, update_refresh_interval: true)
     assert_equal "1s", index.refresh_interval
-    assert_equal "1s", Product.search_index.refresh_interval
+    assert_equal "1s", Product.searchkick_index.refresh_interval
   end
 
   def test_full_resume
@@ -255,64 +306,9 @@ class ReindexTest < Minitest::Test
     assert_equal "wait only available in :async mode", error.message
   end
 
-  def test_callbacks_false
-    Searchkick.callbacks(false) do
-      store_names ["Product A", "Product B"]
-    end
-    assert_search "product", []
-  end
-
-  def test_callbacks_bulk
-    Searchkick.callbacks(:bulk) do
-      store_names ["Product A", "Product B"]
-    end
-    Product.search_index.refresh
-    assert_search "product", ["Product A", "Product B"]
-  end
-
-  def test_callbacks_queue
-    # TODO figure out which earlier test leaves records in index
-    Product.reindex
-
-    reindex_queue = Product.search_index.reindex_queue
-    reindex_queue.clear
-
-    Searchkick.callbacks(:queue) do
-      store_names ["Product A", "Product B"]
-    end
-    Product.search_index.refresh
-    assert_search "product", [], load: false, conversions: false
-    assert_equal 2, reindex_queue.length
-
-    perform_enqueued_jobs do
-      Searchkick::ProcessQueueJob.perform_now(class_name: "Product")
-    end
-    Product.search_index.refresh
-    assert_search "product", ["Product A", "Product B"], load: false
-    assert_equal 0, reindex_queue.length
-
-    Searchkick.callbacks(:queue) do
-      Product.where(name: "Product B").destroy_all
-      Product.create!(name: "Product C")
-    end
-    Product.search_index.refresh
-    assert_search "product", ["Product A", "Product B"], load: false
-    assert_equal 2, reindex_queue.length
-
-    perform_enqueued_jobs do
-      Searchkick::ProcessQueueJob.perform_now(class_name: "Product")
-    end
-    Product.search_index.refresh
-    assert_search "product", ["Product A", "Product C"], load: false
-    assert_equal 0, reindex_queue.length
-
-    # ensure no error with empty queue
-    Searchkick::ProcessQueueJob.perform_now(class_name: "Product")
-  end
-
   def test_object_index
     error = assert_raises(Searchkick::Error) do
-      Product.search_index.reindex(Object.new)
+      Product.searchkick_index.reindex(Object.new)
     end
     assert_equal "Cannot reindex object", error.message
   end
@@ -328,7 +324,7 @@ class ReindexTest < Minitest::Test
   end
 
   def test_both_paths
-    Product.search_index.delete if Product.search_index.exists?
+    Product.searchkick_index.delete if Product.searchkick_index.exists?
     Product.reindex
     Product.reindex # run twice for both index paths
   end